fix #648 Make StepVerifier.withVirtualTime mutally exclusive
This commit adds a lock to the StepVerifier that is only used when
virtual time is activated, resulting in the VirtualTimeScheduler
activation, publisher verification and VTS reset to be mutually
exclusive with other StepVerifier.withVirtualTime verify() invocations.

This prevent different StepVerifier from side-effecting each other, as
the StepVerifier is installed using a global hook.