fix #1105 Lift exposes the original publisher, add adapter for Scannable
Only the filter can be adapted to assume a Scannable, and it is now a
separate liftScannable factory method that clarifies what happens if
the source isn't actually Scannable.

The lifter function itself always deals directly with the original
Publisher.

The javadoc now clarifies that only the lift functions that maintain the
same data type should be used with hooks. Similarly, the onEachOperator
and onLastOperator hooks javadoc both clarify that transformation should
output a Publisher of same data type as the source and that Reactor high
level APIs shouldn't be used.

Also clarify the javadoc