port #1107 Avoid interrupting WorkerTask Future if cancelled race
This commit prevents the undesirable interruption of the WorkerTask's
thread when a cancellation happens within the run() method (eg. a
MonoNext is cancelling upon first onNext downstream).

This would previously lead to a race condition between the cancellation
and the setFuture, which would interrupt if called AFTER the dispose.

This is a backport of #1108 (commit a0b752a), as tracked in #1111