fix #1115 assembly operator have special operatorName, improve Scannable
This commit disambiguates the `Scannable#name()` by aliasing it to
`sequenceName()`. It also adds convenience stream variants to `parents`
and `actuals` that include the current `Scannable`.

`InnerConsumer` subscribers have an `operatorName()` implementation that
returns their simple class name.

Finally, this commit adds a special implementation of operatorName to
the various `xxxOnAssembly` operators and subscribers which returns the
assembly stack trace's first line (trimmed of whitespaces/tabs).