port #1118 onErrorDropped also logs error, MonoCompletionStage drops
This commit makes the MonoCompletionStage drop errors that are thrown
inside the future.whenComplete block (which can happen with fatal
exceptions). Additionally, the default onErrorDropped behavior also logs
the dropped error.

This results in these errors being visible in the logs, where they were
previously swallowed by CompletionFuture.whenComplete.

This is a backport of #1126 (commit 7e29da8), as tracked in #1111