[test] Allow re-run of failed tests (fully filtered test suite, #1139)
This commit prevents test failure when no tests are found due to a too
restrictive filter. This will notably happen when failed tests are
re-run from the IDE, leading to an enumeration of the tests to run that
won't match the includes of additional test tasks like testNG.

Additionally, a small test report is printed for each test task,
giving status of the intermediate states of the global check task.