fix #1115 Add Scannable steps() / stepName()
This commit disambiguates the `Scannable#operatorName()` by aliasing it to
`stepName()`. It also adds a `steps()` method that outputs a `Stream<String>`
of all the stepNames both upstream and downstream (including the current
`Scannable`). `operatorName` is marked as deprecated and to be removed in
3.2.1.

`InnerConsumer` subscribers have a `stepName()` implementation that
returns their simple class name, sources are also identified with the
`SourceProducer` interface with a stepName of "source(CLASS_NAME)".

Finally, this commit adds a special implementation of stepName to
the various `xxxOnAssembly` operators and subscribers which returns the
assembly stack trace's first line (trimmed of whitespaces/tabs).
Such operators are marked with an attribute (`Attr.ACTUAL_METADATA`)
which allows the stepName to be concatenated to the actual operator's
stepName in the `steps()` output.

Most trace-related utility methods have been externalized in a `Traces`
utility class for clarity (for now package-private). Additionally:
 - Checkpoint tracing has also been improved
 - Checkpoint details are included in `steps()` output
 - Trace sanitation removes `org.gradle.*` from traces
 - The `reactor.core.publisher` package is not ignored in test classes

Reviewed in #1140