port #783 Let VirtualTimeScheduler defer advanceTime if queue is empty
This commit changes the way clock advanceTime is performed when the
`VirtualTimeScheduler`'s task queue is empty, deferring the clock move
until a task is scheduled.

This eliminates ordering issues where a scheduling task (like interval)
would happen AFTER the clock was moved (eg. by a `StepVerifier`'s
`thenAwait`), resulting in the interval being scheduled too far away in
the "future" and never happening.

The commit also moves `RaceTestUtils` from core test sourceset to
reactor-test main sourceset and fixes an issue in
`DefaultVerifierSubscriber` where the current expectation is not
refreshed from the script queue after wait tasks have been extracted.

Reviewed-In: #1127
This is a backport of #783 (commit 66051ff), as tracked in #1175