improve #1220 FluxUsingWhen termination/cancel detection + scan
This is a follow up to FluxUsingWhen that improves how termination and
cancellation are detected by the operator (especially for scan's
benefit). It adds a private interface for parents of CommitInner and
RollbackInner so that they can call the parent back to update its
terminated state.

The scanUnsafe operators are now also covered by tests.