[doc] Reflow and reorganize CONTRIBUTING.md guidelines
The contributing guidelines were a bit too cumbersome to read.

Tried to reflow it and push larger blobs of text towards the end of the
document, add a little lighter tone with emojis, emphasize we appreciate
contributions right from the start.

Also took inspiration from the Atom and Ruby on Rails contributing
guidelines mentioned in Github's help pages.