fix #1201 Add utility to instrument Executor backing schedulers...
This commit adds a Schedulers-level API to activate metrics.

Each ExecutorService that backs a Scheduler will get an artificially
generated name "executorId" which avoids overwriting meters when a
Scheduler creates several instrumented ExecutorServices.

The incrementing numbers are computed from a WeakHashMap keyed with the
owning Scheduler instance.

Further Micrometer instrumentation can be implemented if and when
micrometer-metrics/micrometer#1021 is fixed.