fix #1233 UsingWhen can be cancelled early (but loses fusion)
This commit makes UsingWhen cancellable from the main sequence PoV, even
if the resource Publisher has not yet emitted the resource. In that
case, cancellation signal is propagated to the resourcePublisher. If the
resource has been emitted it is propagated to the "main" Publisher that
was derived from the resource.

A side effect of this change is that the operator becomes incompatible
with micro-fusion optimizations: fusion negotiation happens during
onSubscribe, which now happens before the resource is available and thus
before the final main sequence type is known. As this sequence would
have been the one to be fused, negotiation is not possible.