fix #1329 Detect conversion cycle of Mono.from(Flux.from(mono))
This commit allows detection of _some_ cases where a back-and-forth
conversion is applied to a Mono, like `Mono.from(Flux.from(mono))`.

When it is the case (and to the exception of an original Mono that is
Callable), this optimization unwraps the original Mono during the second
conversion, returning the original instance directly.

Callable Monos like `Mono.empty` are not wrapped but resolved and
rewritten as `Flux.empty/error/just`, which makes the optimization
impossible in these cases.