[test] Do not fail HooksTestStaticInit test in IDE (#1322)
The `HooksTestStaticInit` now uses assumptions to be skipped instead
of failing, when executed from the IDE (which can happen eg. when
executing a group of tests).

In the build, if the corresponding Gradle task is invoked there is an
additional check to ensure that test is not skipped accidentally by
some configuration changes.