fix #1347 Remove onDiscard global hook in favor of local only
The global `onDiscard` hook was not provably useful, and could have
unclear caveats when used with libraries that would use the local
hook, as the later would shadow any user-defined global hook. It has
thus been removed entirely.

The `KEY_ON_DISCARD` is back to package private, and we instead offer
a method to add it to a Context: `Operators#enableOnDiscard(Context)`.