[doc] Reorganize docs root (was src)
This commit reorganizes the documentation a bit, removing dead code
and improving the directory hierarchy.

A later goal is to embed marbles in the javadoc jar, by making use of
`doc-files`, but this is provided the total size of the PNG can be
reduced (or a switch to optimized SVG is performed).

In the meantime, this commit makes the following changes:

 - renamed and reorganized /src/ to /docs/
 - api overview and stylesheet now in /docs/api
 - asciidoc now in /docs/asciidoc, images folder correctly copied at
 same level in docsZip
 - moved most marble PNGs from /src/docs/marble to
 /docs/marble/, along with the source graffle file
 - remaining PNGs are used in reference documentation and have been
 moved to /docs/asciidoc/images
 - unused asciidoc stylesheets removed, reference to ../images/ removed,
 golo.css renamed to reactor.css
 - javadoc stylesheet.css now doesn't enforce .marbles to have a 500px
 width, but instead hint at max-width of 800px