fix #1374 Retry-elapsed snippet inconsistency with output in doc
Swapped retry and elapsed operators to obtain the doc output as
advertised.

Having the elapsed() before the retry(1) means that on the start of the
retry attempt, it can only measure the time between retry re-subscribing
to the source and said source emitting its first tick, which is about
250ms. However, the 4th original tick still introduces a "ghost"
delay of 250ms just before that, it is simply not visible in the old
snippet configuration.