fix #1385 Fix refCount(Grace)/Share missing repeat/retry re-subscribes
This commit fixes a bug where `FluxRefCount`/`FluxRefCountGrace` (which
is also accessible through the `FLux#share` API) would miss
re-subscriptions from `repeat` or `retry`. This was due to the state of
connection being reset only AFTER the terminal signal was propagated to
the downstream re-subscribing operator, resulting in the re-subscription
being dismissed by the refcount operator.