fix #1410 Have MonoReduce rely on super.value, nulled out in complete
This commit replaces the redundant `this.result` field with `value`.

Additionally, the field is nulled out upon completion in MonoSubscriber.
Note that subclasses that relied on `value == null` checks to see if
the operator terminated empty now need to also check that the `state`
is STRICTLY INFERIOR to `HAS_REQUEST_HAS_VALUE`.