Add an (untyped) way to choose a default MeterRegistry
This commit adds methods to Metrics that allow choosing a MeterRegistry
that is different from the so far default of _globalRegistry_.

The method is considered "unsafe", because in order to avoid class
loading problems it doesn't expose MeterRegistry as a parameter but
takes a much wider `Object`. Despite that, if Micrometer is on the
classpath, only parameters of type `MeterRegistry` will be taken into
account.