fix #1515 Eagerly avoid logging onNext(QueueSubscription)
New versions of some logging libraries (Log4j2) now swallow _all_
exceptions, including the UnsupportedOperationException that we
previously relied on to detect a case of a type incompatible with "deep
logging".

Now the main culprit, QueueSubscription, is eagerly detected and logged
as a String in safeLog.

We still catch UOE in case other bad log cases happen (with a library
that doesn't swallow the UOE).