fix #1558 Add ToStringConverter and Extractor to StepVerifierOptions
This commit allows users to redefine `String` representation of various
objects that can be included in StepVerifier-produced messages:

 - added 2 new interfaces with implementations in ValueFormatters util
 - StepVerifierOptions can be set up with a single class-matching
 ToStringConverter, along with 0-n Extractors for container values
 - ErrorFormatter is now MessageFormatter
 - it has a more generic format(msg, arg...) method, which is preferred
 to String.format in the DefaultStepVerifier

In options, `getExtractors()` produces a copy of the user-defined
extractors to which it adds defaults at the end. Since only the first
matching extractor applies, this ensures that eg. adding an extractor
for `Integer[]` won't be superseded by the default Object[] extractor.