fix #1666 Use queue.size() for onBackpressureBuffer maxSize evaluation
The queue.offer() can be unreliable, mainly due to Spsc*Queue rounding
the requested capacity up to the nearest power of two, as well as
Queues.get(capacity) rounding up capacity to a minimum of 8...

This commit keeps track of the exact buffer size and compares it to the
queue size() when in bounded mode.

The documentation has been improved to reflect this, as well as the
marble diagram of both variants with a bufferSize and no strategy (the
strategy one being a different implementation that is not impacted).

Tests have been added, and some tests that didn't cover
FluxOnBackpressureBuffer have been moved to the more correct
FluxOnBackpressureStrategyTest case.