ElasticScheduler: improve cache eviction
Currently `ElasticScheduler` uses a queue to manage the cache of executors.
When a task is finished, the free executor gets enqueued to the queue with ttl.
When a new task is scheduled, a cached executor is dequeued from the queue (which has the oldest ttl).

Therefore an executor with the highest chance of getting evicted is removed from the queue for reuse,
which leads to the problem that executors are unable to get evicted if tasks are scheduled at some rate faster than `ttlSeconds`.

This commit replaces queue with deque and reuse executors in LIFO order.
Using deque may have small memory overhead of backward pointer per deque entry,
but it is much smaller than keeping unnecessary thread around.
I assume that performance hit is also small.