fix #1712 Flux.publish().autoConnect(0) should drop early items
This is linked to #1528, the previous fix having introduced a regression
where early connection without subscribers is not handled correctly and
accumulates received items instead of "dropping" them.

This is because, unlike EmitterProcessor, FluxPublish should distinguish
initial connection with empty subscribers array vs empty subscribers
array that is due to removing the last inner (ie. cancelling).

This commit also ensures that disconnectAction, which is synonymous to
cancellation, sets the subscribers array to CANCELLED, not INIT.