[build] Log console output in Gradle ONLY for failed tests on CI
This commit first adds CI detection variable (for TravisCI, CircleCI and
Bamboo).

It also uses that detection to enrich the test logging experience on CI
by logging the console output in Gradle for tests that failed.

This is done by capturing console output and printing it at `afterTest`
if the test result is FAILED.