[doc] Polish 'Suggest Edit' links and asciidoctor config
The links would no longer be rendered with title/class, due to
compatibility mode being used (which I don't think is necessary).

Also simplified the Asciidoctor attributes to only keep the backend
ones in docs.gradle and the presentation ones in index.adoc (notably
the table of content style).

Removed the `linkattrs` attribute since it is no longer needed to
interpret roles in links.