fix #1739 Enforce a name for Schedulers.fromExecutorService
This avoids relying on the underlying ExecutorService's toString, at
least for the purpose of calling `scan(Attr.NAME)`.

Previously, relying on the backing ExecutorService#toString method
was not ideal as this generally depends on the state of the executor,
which is not ideal for most usages of `NAME` and especially not for
metrics (where NAME is used a a differentiating tag for the meters).

There is an additional issue of NAME being used for metrics decorator,
which is invoked during construction. Other schedulers use fields that
are initiliazed early in the constructor, but fromExecutorService used
the executor toString, which obviously cannot possibly be initialized.
Adding the name field and initializing it early helps again here.