fix #1783 Do not drop errors in MonoCacheTime, better doc (#1784)
This commit avoids onErrorDropped in the path where TTL function throws
and the original signal was an error. Since the default behavior of
dropping is to throw, this would break the operator: no state reset
nor exception propagation to downstream would occur.

Additionally, the case of TTL 0 is now explicitly handled and avoids
going through the Scheduler, immediately resetting the state.

This is now better documented, and the TTL 0 approach gets more focus
in the javadoc (as throwing TTL function is more of an edge case).