fix #1778 Deprecate BiFunction scheduler decorators, change metrics
Since the BiFunction took the Scheduler reference, but was mostly
applied from inside said Scheduler constructor, it would imply dealing
with a not fully constructed object.

Switching to a decorator interface that takes two Strings (scheduler
type and soft id) allows for a more reliable (yet more restricted)
mechanism.

The two approaches cannot be sharing the same underlying implementation,
which has the unfortunate consequence that implementors MUST apply the
two sets of decorators explicitly.

The Schedulers.enableMetrics() also need to change in the new approach,
removing the capacity to detect executors that are attached to a common
instance, and instead aggregating all similarly configured Schedulers
under their common schedulerId. On the other hand, the meters can now
be tagged with the more generic Scheduler type (eg. all PARALLELs).