fix #1797 Explicitly reject null keys/values in Context constructors
This commit better enforces non-null nature of the various Context
implementations by throwing a NullPointerException whenever a null key
or null value is passed to one of the implementations' constructors.

This aligns with the `put` behavior, which already throws NPE.

To that end, ContextN now extends ConcurrentHashMap: this flavor
naturally rejects nulls, contrary to HashMap and LinkedHashMap.