fix #1745 Keep transform() and rename compose() to transformDeferred()
This commit mostly reverts d7b2271 and reworks the fix into keeping
the `transform()` method as it originally was, focusing on `compose()`.

The later is renamed to `transformDeferred`, conveying its relationship
to `defer()` in the sense that both apply per-subscriber. This allows
the two APIs to be easily discoverable in code completion while keeping
the most frequently version untouched.