fix #1451 Dematerialize needs extra request(1) for terminal signals...
This commit reworks dematerialize so that it maps 1 request to 1 Signal
from upstream, immediately propagating that decoded signal.

As a consequence, when requested one by one, dematerialize now requires
one more request for the terminal signal to be pulled and propagated.

This also means that the requested signals won't be delayed until the
upstream emits the next signal (which was both problematic in terms of
timing consistency and in terms of getting the last `onNext` in the case
of a never ending source).