see #1556 Macro-fuse doOn* operators, but let all doOnError hit
This commit activates macro-fusion on FluxPeek(Fuseable) operators
(doOnSubscribe, doOnNext, doOnComplete, doOnError, doOnRequest,
doOnCancel, doOnTerminate and doAfterTerminate).

As a consequence of combining similar Consumers/Runnable, the first
lambda to throw will prevent other handlers to execute. Since this was
most undesirable for `doOnError` (where we expect each lambda to receive
the previous error suppressed by the the new error), the combination of
errorHandlers is a bit more complex with an integrated try/catch.