see #1556 Macro-fuse some Mono#doOn* operators, let all doOnError hit
This commit activates macro-fusion on MonoPeek(Fuseable) operators
(doOnSubscribe, doOnNext, doOnError, doOnRequest,
doOnCancel and doOnTerminate). Contrary to the Flux counterpart, this
doesn't cover `doAfterTerminate`. It doesn't cover `doOnSuccess`,
`doOnSuccessOrError` and `doAfterSuccessOrError` either.

As a consequence of combining similar Consumers/Runnable, the first
lambda to throw will prevent other handlers to execute. Since this was
most undesirable for `doOnError` (where we expect each lambda to receive
the previous error suppressed by the the new error), the combination of
errorHandlers is a bit more complex with an integrated try/catch.