[build] Polish benchmarks build, remove jmh plugin
This commit removes the jmh-gradle-plugin in favor of self-defined
jmh task, following difficulties making the plugin alternative work in
both the IntelliJ IDE and the CLI (pure Gradle).

In order to make things work, the `main` sourceSet is used, which is
more naturally supported by IDEs. JMH annotation processing is applied
through the recent `annotationProcessor` dependency keyword in Gradle.