Polish jmh task include parameter: auto-add wildcards
Now `-Pinclude` automatically wraps the parameter in wildcards, and in
order to specify the full regexp explicitly, one can use `-PfullInclude`
instead. The former name of the include parameter, `jmh`, caused
confusion with the task of the same name and thus would break execution
when no such parameter was specified.