fix #1706 Add BufferUntilChanged operator
This commit lets predicates used in `bufferUntil` have state even when needing
cleanup, simply by virtue of being `Disposable`.

The untilChanged predicate is pre-implemented and offered as an alias,
`bufferUntilChanged`, which works similarly to `distinctUntilChanged` in regard
to which elements get grouped into buffer (using both key extractor and key
comparator). The `lastKey` internal state is cleared up upon termination or
cancellation of the sequence.