fix #1752 Make doOnTerminate for Mono same as doOnSuccess
This means that the handler is invoked during processing of the `onNext`
signal.

This commit also slightly rework MonoPeekTerminal to make the onSuccess
and onError handlers canonical, instead of the (successOrError) one.

Finally, it makes `doOnError` directly use MonoPeekTerminal instead of
MonoPeek, and modifies the marble diagrams (adding a convention for the
null value / empty set).

Relates-to: #1853