cherry-pick #1872 Avoid classloader leak due to static throwables
When the stacktrace of a `Throwable` is filled, the JVM transparently
adds a `backtrace` field that references classes from the stacktrace.
In multi-classloader environments, this can lead to leaks when the
`Throwable` (of subclass of) is static.

This commit changes the few occurrences of static exceptions in Reactor
so that they each use a variant that doesn't fill the stack trace.

TERMINATED uses a general-purpose StaticThrowable which means that it
can reference the super constructor with boolean flags disabling
exception suppression and stack trace filling, without having to
override the fillInStackTrace method.

Reviewed-in: #1876