see #1804 Enforce 100K task cap on Schedulers.boundedElastic()
While introducing the singleton version of BoundedElasticScheduler,
we might as well make it truly bounded, including in the number of task
submissions it can "absorb" during a spike. Used to be unbounded, now
bounded to 100K tasks.

This is also more likely to uncover problematic scheduling of blocking
tasks. Users switching from `elastic()` to `boundedElastic()` and
seeing `RejectedExecutionException`s should challenge why their Reactor
app schedules so many blocking tasks, and if still legitimate and
necessary, use a `newBoundedElastic` instance that is better tuned to
their workload.

Reviewed-in: #1900