fix #1913 Offer alternative to `expectNoEvent(d).thenCancel()`
The `expectTimeout(d)` method is a preferred alternative to the
`expectNoEvent(d).thenCancel()` combination. It should do a better job
at avoiding false positives, and is more expressive of the expectation
that a Publisher "never ends" (although it is impossible to meaningfully
assert that without waiting for the heat death of the universe).