fix #1945 Avoid unnecessary wrapping of exceptions in scalar flatMapping
Due to the usage of Operators.onNextPollError for the purpose of error
mode with scalar sources, which returns a RuntimeException, the original
errors can be wrapped in a ReactiveException (provided they're not a
RuntimeException already).

This is problematic, since this occurs even if no error mode strategy
is in place.

This change introduces a new Operators.onNextError variant that doesn't
wrap the exception nor cancels a Subscription. It merely returns the
original error to be passed to onError.