fix #1949 Avoid race between request handler setup and first requests
When the request handler is being set up in the create/push lambda,
one or several concurrent request can occur. In that case, the
consumer might miss requests.

This commit fixes the situation by tracking requests made before the
consumer could be registered, with a special flag of `-1` when it is
about to be registered. In the later case, in `#request(long)`, we
wait until we see the `Consumer` and we directly pass the request
amount to it.