fix #1962 Avoid extending Entry in Context1
The motivation for this change is that this drives two implementations
of `Context#stream()` to use the `Context1` as an immutable map entry.

In turn, this proves problematic when the `stream()` is converted to a
string representation (eg. by loggers or assertions): the stream entries
use `Context1{k=v}` format, which is fine in a standalone toString but
less so in a representation of the whole stream.

AbstractMap.SimpleImmutableEntry can be leveraged instead, and has a
proper `key=value` string representation by default.

Also added tests around putAll-replaces case.