fix #1970 Avoid creating Map.Entry in Context.putAll, ContextN delegates
This commit introduces `AbstractContext` as a common base for all
reactor-provided `Context` implementations.

This allows to invert the control of the putAll method, avoiding
external iteration via the `stream()` and thus wasteful production of
`Map.Entry` objects.

As a consequence, `ContextN` cannot extend LinkedHashMap anymore and
thus delegates to it instead, so it can extend AbstractContext.