Follow up to #1979 to avoid retryWhen ambiguity (#2079)
Retry is now an abstract class, which lift the lambda ambiguity
(since this disallows lambda usage). To help expressing simple
Retry strategies as lambdas, a Retry.from(Function) is provided.

A couple of factories have been added:
 - indefinitely() is a RetrySpec configured to always immediately
retry (but can be further configured)
 - fixedDelay(maxAttempts, duration) is a RetryBackoffSpec
configured with a minBackoff equal to maxBackoff and no
jitter, effectively emulating a fixed delay retry

Additionally, a few polish are applied to the original PR:
 - javadocs have been improved
 - marble diagrams have been added to the Retry factories
 - for RetryBackoffSpec, one can now reset the default Scheduler
by passing null. That default is also lazily resolved, making it
compatible with StepVerifier#withVirtualTime